feature: make csv file that collects all FSO px cubes and metadata
The csv file of px cube ids is generated first.
Afterwards it is enriched in batches of a step size that can be selected by
dimensions and other table metadata. The enrichment needs to be done in steps
because otherwise the api is overloaded and crashes. Between the batches the
the script pauses for some seconds before fetching the next batch